# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Language Learning Platform API
  version: 0.1.0
servers:
  - url: https://api.example.com
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  env:
                    type: string

  /auth/me:
    post:
      summary: Exchange WeChat code for user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
              required: [code]
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  featureFlags:
                    type: array
                    items:
                      type: string

  /content:
    get:
      summary: List learning content
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: type
          schema: { type: string, enum: [sentence, dialog, word] }
        - in: query
          name: level
          schema: { type: string }
      responses:
        '200':
          description: Content page
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContentItem'
                  total:
                    type: integer

  /content/{id}:
    get:
      summary: Get content by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Content item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentItem'

  /study/progress:
    post:
      summary: Save study progress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contentId: { type: string }
                secs: { type: integer }
                step: { type: string, enum: [new, learning, review] }
              required: [contentId, secs, step]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'

  /vocab:
    get:
      summary: List vocab
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Vocab page
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/VocabItem' }
                  total:
                    type: integer
    post:
      summary: Add vocab item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VocabAddRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }

  /vocab/{word}:
    delete:
      summary: Remove vocab item
      parameters:
        - in: path
          name: word
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }

  /vocab/review:
    post:
      summary: Review a vocab item with SRS rating
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VocabReviewRequest'
      responses:
        '200':
          description: Updated vocab item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VocabReviewResponse'

  /vocab/due:
    get:
      summary: List vocab items due by a timestamp
      parameters:
        - in: query
          name: before
          schema: { type: string, description: ISO8601 timestamp }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Due vocab page
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/VocabItem' }
                  total: { type: integer }

  /plan/stats:
    get:
      summary: Get today's seconds and current streak
      responses:
        '200':
          description: Plan stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanStatsResponse'

  /events:
    post:
      summary: Track an analytics event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event: { type: string }
                props: { type: object }
                ts: { type: string, description: ISO8601 timestamp }
              required: [event]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'

  /import:
    post:
      summary: Import from URL or text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportRequest'
      responses:
        '200':
          description: Parsed content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResponse'

  /export/longshot:
    post:
      summary: Generate long image export
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportLongshotRequest'
      responses:
        '200':
          description: Export URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportLongshotResponse'

  /billing/intent:
    post:
      summary: Create upgrade intention
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIntentRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }

  /wishlist:
    post:
      summary: Submit wishlist contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WishlistRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }

  /cron/daily-digest:
    post:
      summary: Trigger daily digest
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                dryRun: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }

components:
  schemas:
    Ok:
      type: object
      properties:
        ok: { type: boolean }
    User:
      type: object
      properties:
        openid: { type: string }
        nick: { type: string, nullable: true }
        avatar: { type: string, nullable: true }
        pro_until: { type: string, nullable: true }
      required: [openid]

    ContentItem:
      type: object
      properties:
        id: { type: string }
        type: { type: string, enum: [sentence, dialog, word] }
        tags: { type: array, items: { type: string } }
        level: { type: string, nullable: true }
        text: { type: string }
        audio_url: { type: string, nullable: true }
        segments: { type: array, items: { type: string } }
        segments_audio: { type: array, items: { type: string }, nullable: true }
      required: [id, type, text]
    VocabItem:
      type: object
      properties:
        word: { type: string }
        meaning: { type: string, nullable: true }
        phonetic: { type: string, nullable: true }
        srs_level: { type: integer, nullable: true }
        next_review_at: { type: string, nullable: true }
      required: [word]

    VocabAddRequest:
      type: object
      properties:
        word: { type: string }
        meaning: { type: string }
        phonetic: { type: string }
      required: [word]

    ImportRequest:
      type: object
      properties:
        type: { type: string, enum: [text, url] }
        payload: { type: string }
      required: [type, payload]

    ImportResponse:
      type: object
      properties:
        id: { type: string }
        type: { type: string, enum: [text, url] }
        segments: { type: array, items: { type: string } }
      required: [id, type, segments]

    ExportLongshotRequest:
      type: object
      properties:
        contentId: { type: string }
      required: [contentId]

    ExportLongshotResponse:
      type: object
      properties:
        url: { type: string }
      required: [url]

    CreateIntentRequest:
      type: object
      properties:
        plan: { type: string }
        price: { type: number }
      required: [plan, price]

    WishlistRequest:
      type: object
      properties:
        contactType: { type: string }
        contact: { type: string }
        plan: { type: string }
        price_point: { type: number }
      required: [contactType, contact]
    VocabReviewRequest:
      type: object
      properties:
        word: { type: string }
        rating: { type: string, enum: [again, hard, good, easy] }
      required: [word, rating]

    VocabReviewResponse:
      type: object
      properties:
        item: { $ref: '#/components/schemas/VocabItem' }
      required: [item]
    PlanStatsResponse:
      type: object
      properties:
        secsToday: { type: integer }
        streak: { type: integer }
      required: [secsToday, streak]